+++ shared-helpers/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.5](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.3.4...@bloom-housing/shared-helpers@7.3.5) (2023-02-07)
+
+**Note:** Version bump only for package @bloom-housing/shared-helpers
+
+
+
+
+
+## [7.3.4](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.3.3...@bloom-housing/shared-helpers@7.3.4) (2023-02-07)
+
+**Note:** Version bump only for package @bloom-housing/shared-helpers
+
+
+
+
+
+## [7.3.3](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.3.2...@bloom-housing/shared-helpers@7.3.3) (2023-02-06)
+
+**Note:** Version bump only for package @bloom-housing/shared-helpers
+
+
+
+
+
+## [7.3.2](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.3.1...@bloom-housing/shared-helpers@7.3.2) (2023-02-01)
+
+
+### Bug Fixes
+
+* add main content id ([#3299](https://github.com/bloom-housing/bloom/issues/3299)) ([062335b](https://github.com/bloom-housing/bloom/commit/062335bd0a1a7af2cc9e7f2bc9a44dad6163d3d9))
+
+
+
+
+
+## [7.3.1](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.3.0...@bloom-housing/shared-helpers@7.3.1) (2023-02-01)
+
+**Note:** Version bump only for package @bloom-housing/shared-helpers
+
+
+
+
+
 # [7.3.0](https://github.com/bloom-housing/bloom/compare/@bloom-housing/shared-helpers@7.2.0...@bloom-housing/shared-helpers@7.3.0) (2023-01-26)